chore(uploader): move records metadata uplaoder to separate directory
