+++ uploader-records-metadata/lib/entities/collection.js
