+++ uploader-records-metadata/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "audio-interface-of-internet-archive/uploader-records-metadata",
+  "description": "Audio Interface of Internet Archive (uploader-records-metadata)",
+  "version": "2.5.11",
+  "private": true,
+  "author": "Internet Archive",
+  "license": "AGPL-3.0",
+  "engines": {
+    "node": "8"
+  },
+  "scripts": {
+    "lint": "eslint ./",
+    "uploader": "node ./uploader",
+    "mocha": "mocha --opts ./tests/mocha.opts",
+    "start": "firebase serve --only functions:assistant",
+    "start:alexa": "bst proxy lambda index-alexa",
+    "deploy": "firebase deploy --only functions:assistant",
+    "test": "npm run lint && npm run mocha"
+  },
+  "dependencies": {
+    "actions-on-google": "^2.9.1",
+    "alexa-sdk": "^1.0.25",
+    "ask-sdk": "^2.6.0",
+    "axios": "^0.19.0",
+    "bespoken-tools": "^2.3.38",
+    "child-process-promise": "^2.2.1",
+    "dashbot": "^10.4.1",
+    "debug": "^4.1.1",
+    "firebase-admin": "^8.3.0",
+    "firebase-functions": "^3.2.0",
+    "glob": "^7.1.4",
+    "hirestime": "^3.2.1",
+    "lodash": "^4.17.15",
+    "mathjs": "^5.10.3",
+    "moment": "^2.24.0",
+    "mustache": "^2.3.2",
+    "node-fetch": "^2.6.0",
+    "object.entries": "^1.1.0",
+    "raven": "^2.6.4",
+    "replaceall": "^0.1.6",
+    "uuid": "^3.3.2"
+  },
+  "devDependencies": {
+    "axios-mock-adapter": "^1.17.0",
+    "chai": "^4.2.0",
+    "eslint": "^5.16.0",
+    "eslint-config-semistandard": "^13.0.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-chai-friendly": "^0.4.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-mocha": "^5.3.0",
+    "eslint-plugin-node": "^8.0.1",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "js-yaml": "^3.13.1",
+    "mocha": "^6.2.0",
+    "rewire": "^4.0.1",
+    "sinon": "^7.3.2",
+    "sinon-chai": "^3.3.0",
+    "virtual-alexa": "^0.7.2"
+  }
+}
